Fixed : cURL uploader Fixed : isSectionEnabled check on view channel

+++ upload/actions/file_downloader.php
@@ -255,7 +255,10 @@ sleep(10);
 $details =  file_get_contents($log_file);
 $details = json_decode($details,true);
 $Upload->add_conversion_queue($details['file_name']);
+
+if(file_exists($log_file))
 unlink($log_file);
+if(file_exists($dummy_file))
 unlink($dummy_file);
 $quick_conv = config('quick_conv');
 $use_crons = config('use_crons');
@@ -263,6 +266,8 @@ $use_crons = config('use_crons');
 
 //Inserting data
 $title 	= urldecode(mysql_clean(getName($file)));
+$title = $title ? $title : "Untitled";
+
 $vidDetails = array
 (
 	'title' => $title,
